+++ lib/src/better_feedback.dart
@@ -1,3 +1,4 @@
+import 'package:feedback/feedback.dart';
 import 'package:feedback/src/feedback_controller.dart';
 import 'package:feedback/src/feedback_functions.dart';
 import 'package:feedback/src/feedback_widget.dart';
@@ -15,7 +16,7 @@ class BetterFeedback extends StatefulWidget {
     this.theme,
     this.localizationsDelegates,
     this.localeOverride,
-    this.defaultNavigate = true,
+    this.mode = FeedbackMode.navigate,
   }) : super(key: key);
 
   /// The application to wrap, typically a [MaterialApp].
@@ -37,10 +38,10 @@ class BetterFeedback extends StatefulWidget {
   /// If no platform default locale exists, english is used.
   final Locale? localeOverride;
 
-  /// Set the default controlling mode when launching feedback
-  /// By default it will use start with the navigation option
-  /// Set to false to launch in drawing mode
-  final bool defaultNavigate;
+  /// Set the default mode when launching feedback.
+  /// By default it will allow the user to navigate.
+  /// See [FeedbackMode] for other options.
+  final FeedbackMode mode;
 
   /// Call `BetterFeedback.of(context)` to get an instance of
   /// [FeedbackData] on which you can call `.show()` or `.hide()`
@@ -85,7 +86,7 @@ class _BetterFeedbackState extends State<BetterFeedback> {
                 child: widget.child,
                 isFeedbackVisible: feedbackVisible,
                 drawColors: FeedbackTheme.of(context).drawColors,
-                defaultNavigate: widget.defaultNavigate,
+                mode: widget.mode,
               );
             },
           ),